requirements.txt
requests==2.32.3
openai==1.68.0
python-dotenv==1.1.0
